+++ lib/access/index.ts
@@ -1,0 +1,2 @@
+export * from "./permission-catalog";
+export * from "./system-roles";